Relocating the web socket wrapper.
