+++ test/clients/automower/automowerEventStreamClient.spec.ts
@@ -3,8 +3,8 @@ import { It, Mock, Times } from 'moq.ts';
 import * as constants from '../../../src/settings';
 
 import { AutomowerEvent, AutomowerEventTypes, ConnectedEvent, ErrorEvent } from '../../../src/clients/automower/automowerEventStreamClient';
-import { WebSocketWrapper } from '../../../src/clients/primitives/webSocketWrapper';
 import { PlatformLogger } from '../../../src/diagnostics/platformLogger';
+import { WebSocketWrapper } from '../../../src/primitives/webSocketWrapper';
 import { AutomowerEventStreamClientImplSpy } from './automowerEventStreamClientImplSpy';
 
 describe('AutomowerEventStreamClientImpl', () => {
